Clens a few things up.
