+++ src/qforte/abc/__init__.py
@@ -1,2 +1,5 @@
 from .algorithm import *
 from .qsdabc import *
+
+from .vqeabc import *
+from .uccvqeabc import *
