+++ src/qforte/abc/algorithm.py
@@ -2,6 +2,52 @@ from abc import ABC, abstractmethod
 from qforte.utils.state_prep import *
 
 class Algorithm(ABC):
+    """
+    Attributes
+    ----------
+    _ref : list
+        The set of 1s and 0s indicating the initial quantum state.
+
+    _nqb : int
+        The number of qubits the calculation empolys.
+
+    _qb_ham : QuantumOperator
+        The operator to be measured (usually the Hamiltonain), mapped to a
+        qubit representation.
+
+    _fast : bool
+        Whether or not to use a faster version of the algorithm that bypasses
+        measurment (unphysical for quantum computer).
+
+    _trotter_order : int
+        The Trotter order to use for exponentiated operators.
+        (exact in the infinte limit).
+
+    _trotter_number : int
+        The Trotter number (or the number of trotter steps)
+        to use for exponentiated operators.
+        (exact in the infinte limit).
+
+    _Egs : float
+        The final ground state energy value.
+
+    _Umaxdepth : QuantumCircuit
+        The deepest circuit used during any part of the algorithm.
+
+    _n_ham_measurements : int
+        The total number of times the energy was evaluated via
+        measurement of the Hamiltoanin
+
+
+
+    Methods
+    -------
+    build_Uprep()
+        Returns a QuantumCircuit object corresponding to the state preparation
+        circuit reference state (usually a small product of X gates).
+
+
+    """
 
     def __init__(self,
                  system,
@@ -69,9 +115,8 @@ class Algorithm(ABC):
 #         if self._Umaxdepth is None:
 #             raise NotImplementedError('Concrete Algorithm class must define self._Umaxdepth attribute.')
 
-#         if self._tot_Nmeasurements is None:
-#             raise NotImplementedError('Concrete Algorithm class must define self._tot_Nmeasurements attribute.')
+#         if self._n_ham_measurements is None:
+#             raise NotImplementedError('Concrete Algorithm class must define self._n_ham_measurements attribute.')
 
-#         if self._tot_Npreps is None:
-#             raise NotImplementedError('Concrete Algorithm class must define self._tot_Npreps attribute.')
-        
+#         if self._n_ham_measurements is None:
+#             raise NotImplementedError('Concrete Algorithm class must define self._n_ham_measurements attribute.')
