+++ src/qforte/ucc/__init__.py
@@ -1,3 +1,5 @@
 from .ucc_helpers import *
 from .nucc import *
 from .adapt import *
+from .newadaptvqe import *
+from .newuccsdvqe import *
