+++ src/qforte/ucc/newuccsdvqe.py
@@ -185,8 +185,7 @@ class UCCSDVQE(UCCVQE):
             optimizer='BFGS',
             use_analytic_grad = True):
 
-        # TODO (cleanup): add option to
-
+        # TODO (cleanup): add option to pre populate cluster amps.
         self._opt_thresh = opt_thresh
         self._opt_maxiter = opt_maxiter
         self._use_analytic_grad = use_analytic_grad
@@ -280,96 +279,6 @@ class UCCSDVQE(UCCVQE):
         print('Total number of comutator measurements:      ', self.get_num_comut_measurements())
 
     # Define VQE abstract methods.
-
-    # TODO (opt major): write a C function that prepares this super efficiently
-    def build_Uvqc(self, params=None):
-        """ This function returns the QuantumCircuit object built
-        from the appropiate ampltudes (tops)
-
-        Parameters
-        ----------
-        params : list
-            A lsit of parameters define the variational degress of freedom in
-            the state perparation circuit Uvqc.
-        """
-        sq_ops = []
-        for j, pool_idx in enumerate(self._tops):
-            appended_term = copy.deepcopy(self._pool[pool_idx])
-            for l in range(len(appended_term)):
-                if params is None:
-                    appended_term[l][1] *= self._tamps[j]
-                else:
-                    appended_term[l][1] *= params[j]
-                sq_ops.append(appended_term[l])
-
-        Uorg = get_ucc_jw_organizer(sq_ops, already_anti_herm=True)
-        A = organizer_to_circuit(Uorg)
-
-        U, phase1 = trotterize(A, trotter_number=self._trotter_number)
-        Uvqc = qforte.QuantumCircuit()
-        Uvqc.add_circuit(self._Uprep)
-
-        Uvqc.add_circuit(U)
-        if phase1 != 1.0 + 0.0j:
-            raise ValueError("Encountered phase change, phase not equal to (1.0 + 0.0i)")
-
-        return Uvqc
-
-    def measure_gradient(self, HAm, Ucirc):
-        """
-        Parameters
-        ----------
-        HAm : QuantumOperator
-            The comutator to measure.
-
-        Ucirc : QuantumCircuit
-            The state preparation circuit.
-        """
-        # TODO (cleanup): remove N_samples as argument (implement variance based thresh)
-        if self._fast:
-            myQC = qforte.QuantumComputer(self._nqb)
-            myQC.apply_circuit(Ucirc)
-            val = 2*np.real(myQC.direct_op_exp_val(HAm))
-        else:
-            # TODO (cleanup): remove N_samples as argument (implement variance based thresh)
-            Exp = qforte.Experiment(self._nqb, Ucirc, HAm, 1000)
-            empty_params = []
-            val = 2*Exp.perfect_experimental_avg(empty_params)
-
-        assert(np.isclose(np.imag(val),0.0))
-        return val
-
-    def measure_energy(self, Ucirc):
-        """
-        Parameters
-        ----------
-        Ucirc : QuantumCircuit
-            The state preparation circuit.
-        """
-        if self._fast:
-            myQC = qforte.QuantumComputer(self._nqb)
-            myQC.apply_circuit(Ucirc)
-            val = np.real(myQC.direct_op_exp_val(self._qb_ham))
-        else:
-            Exp = qforte.Experiment(self._nqb, Ucirc, self._qb_ham, 1000)
-            empty_params = []
-            val = Exp.perfect_experimental_avg(empty_params)
-
-        assert(np.isclose(np.imag(val),0.0))
-        return val
-
-    def energy_feval(self, params):
-        Ucirc = self.build_Uvqc(params=params)
-        return self.measure_energy(Ucirc)
-
-    def gradient_ary_feval(self, params):
-        Uvqc = self.build_Uvqc(params=params)
-        grad_lst = []
-        for m in self._tops:
-            grad_lst.append(self.measure_gradient(self._comutator_pool[m], Uvqc))
-
-        return np.asarray(grad_lst)
-
     def solve(self):
         """
         Parameters
@@ -415,19 +324,6 @@ class UCCSDVQE(UCCVQE):
             self._Egs = res.fun
             self._final_result = res
             self._tamps = list(res.x)
-
-    # Define UCCSD-VQE methods.
-
-    def fill_comutator_pool(self):
-        print('\n\n==> Building comutator pool for gradient measurement.')
-        # TODO (opt): Perhaps after moving operator handling to to C side.
-        for i in range(len(self._pool)):
-            Am_org = get_ucc_jw_organizer(self._pool[i], already_anti_herm=True)
-            H_org = circuit_to_organizer(self._qb_ham)
-            HAm_org = join_organizers(H_org, Am_org)
-            HAm = organizer_to_circuit(HAm_org) # actually returns a single-term QuantumOperator
-            self._comutator_pool.append(HAm)
-        print('==> Comutator pool construction complete.')
 
     def initialize_ansatz(self):
         for l in range(len(self._pool)):
